requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
web3>=7.9.0